+++ modules/javafx.graphics/src/main/java/com/sun/javafx/tk/quantum/ZoomGestureRecognizer.java
@@ -188,7 +188,7 @@ class ZoomGestureRecognizer implements GestureRecognizer {
                 sendZoomFinishedEvent();
             }
             if (ZOOM_INERTIA_ENABLED && (state == ZoomRecognitionState.PRE_INERTIA || state == ZoomRecognitionState.ACTIVE)) {
-                double timeFromLastZoom = (time - zoomStartTime) / 1000000;
+                double timeFromLastZoom = ((double)time - zoomStartTime) / 1000000;
                 if (initialInertiaZoomVelocity != 0 && timeFromLastZoom < 200) {
                     state = ZoomRecognitionState.INERTIA;
                     // activate inertia
@@ -275,7 +275,7 @@ class ZoomGestureRecognizer implements GestureRecognizer {
                         totalZoomFactor *= zoomFactor;
                         sendZoomEvent(false);
                         distanceReference = currentDistance;
-                        double timePassed = (time - zoomStartTime) / 1000000000;
+                        double timePassed = ((double)time - zoomStartTime) / 1000000000;
                         if (timePassed > 1e-4) {
                             initialInertiaZoomVelocity = (totalZoomFactor - prevTotalZoomFactor) / timePassed;
                             zoomStartTime = time;
